requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
google-genai==0.3.0
